Create delaycreating.ino
How to create delay on Arduino basics? Code to create delay on Arduino using basic clock principle ? Self explanatory code to create delay in Arduino This i/Atmega328p MC. Good for educative purpose and for working too. This is one of the two methods that are used to create delay, this is the simple one, another one is more reliable. Same concept will work for any MC or MP. But check MIPS and MHz relation of that chip through data sheet.